Move radix_tree to a separate module
